requirements.txt
Flask==3.1.1
discord.py==2.4.0
python-dotenv==1.0.1
pymongo==4.6.3